snapcraft: Compile matterhorn from upstream sources
Use upstream develop branch as base source, so that we can have a proper
edge snap channel open with all the upstream git changes.

New releases will also be included in the same branch but the version
will not include the -g${SHA} suffix so will be easy to spot.

Since we use core18, and given that the `cabal` provided by bionic isn't
enough to compile matterhorn, I'm using the same PPA that upstream is
using to get both cabal and ghc.

Alternatively it could be possible to install new cabal with just `cabal
install cabal-install` but this would lead to compiling more components
that would slow-down the building time, and this is quite unneeded since
upstream-trusted build-depends binaries exist.

The ppa usage can be reverted once we'll be in core20.

Compiling this with snapcraft + multipass requires more RAM than you're
generally given, so `SNAPCRAFT_BUILD_ENVIRONMENT_MEMORY=4G` or similar env
variable should be passed when the multipass VM is created.